fix: correct mdt data updates

+++ sv_mdt.lua
@@ -176,12 +176,14 @@ AddEventHandler("bucky_mdt:saveOffenderChanges", function(charidentifier, change
 
 		if changes.convictions ~= nil then
 			for conviction, amount in pairs(changes.convictions) do	
-				exports.oxmysql:execute('UPDATE `user_convictions` SET `count` = ? WHERE `char_id` = ? AND `offense` = ?', {charidentifier, amount, conviction})
+				exports.oxmysql:execute('UPDATE `user_convictions` SET `count` = ? WHERE `char_id` = ? AND `offense` = ?', {amount, charidentifier, conviction})
 			end
 		end
 
-		for i = 1, #changes.convictions_removed do
-			exports.oxmysql:execute('DELETE FROM `user_convictions` WHERE `char_id` = ? AND `offense` = ?', {charidentifier, changes.convictions_removed[i]})
+		if changes.convictions_removed then
+			for i = 1, #changes.convictions_removed do
+				exports.oxmysql:execute('DELETE FROM `user_convictions` WHERE `char_id` = ? AND `offense` = ?', {charidentifier, changes.convictions_removed[i]})
+			end
 		end
 
 		TriggerClientEvent("bucky_mdt:sendNotification", usource, Config.Notify['1'])
@@ -190,7 +192,7 @@ end)
 
 RegisterServerEvent("bucky_mdt:saveReportChanges")
 AddEventHandler("bucky_mdt:saveReportChanges", function(data)
-	exports.oxmysql:execute('UPDATE `mdt_reports` SET `title` = ?, `incident` = ? WHERE `id` = ?', {data.id, data.title, data.incident})
+	exports.oxmysql:execute('UPDATE `mdt_reports` SET `title` = ?, `incident` = ? WHERE `id` = ?', {data.title, data.incident, data.id})
 	TriggerClientEvent("bucky_mdt:sendNotification", source, Config.Notify['2'])
 end)
 
@@ -224,7 +226,7 @@ AddEventHandler("bucky_mdt:submitNewReport", function(data)
 	for offense, count in pairs(data.charges) do
 		exports.ghmattimysql:execute('SELECT * FROM `user_convictions` WHERE `offense` = ? AND `char_id` = ?', {offense, data.char_id}, function(result)
 			if result[1] then
-				exports.oxmysql:execute('UPDATE `user_convictions` SET `count` = ? WHERE `offense` = ? AND `char_id` = ?', {data.char_id, offense, count + 1})
+				exports.oxmysql:execute('UPDATE `user_convictions` SET `count` = ? WHERE `offense` = ? AND `char_id` = ?', {count + 1, offense, data.char_id})
 			else
 				exports.oxmysql:insert('INSERT INTO `user_convictions` (`char_id`, `offense`, `count`) VALUES (?, ?, ?)', {data.char_id, offense, count})
 			end
